+++ crates/rust-eigenda-v2-common/src/core/check_da_cert_status.rs
